gateway/backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
python-dotenv>=1.0.0
pytest>=8.0.0
anyio>=4.0.0
httpx>=0.27.0
